move logic into separate dir
