+++ src/logic/getBoardNodesFromFlatIndex.js
