+++ src/logic/tallyScore.js
