+++ src/logic/tallyScore.test.js
