+++ src/logic/validDropQ.js
